checkpoint refactored for multiple users

+++ src/components/ad.js
@@ -2,13 +2,19 @@ import React, { Component } from 'react';
 
 class Ad extends Component {
 
+  deleteButton = () => {
+    if (this.props.handleDelete){
+    return <button onClick = {() => {this.props.handleDelete(this.props.ad_id)}}>delete</button>
+    }
+  }
+  
   render(){
   return (
    <div>
        <h1>ad</h1>
         <p>food name: {this.props.ad.food_name}</p>
         <p>post code: {this.props.ad.postcode}</p>
-        <button onClick = {() => {this.props.handleDelete(this.props.ad_id)}}>delete</button>
+        {this.deleteButton()}
        <h1></h1>
    </div>
